delete

+++ MiaClient/FrmBeheerPrioriteit.cs
@@ -158,7 +158,11 @@ namespace MiaClient
         private void btnVerwijderen_Click(object sender, EventArgs e)
         {
             
-        
+            if (LstPrioriteiten.SelectedItem== null)
+            {
+                MessageBox.Show("Er is geen financieringstype geselecteerd om te verwijderen.","MIA",MessageBoxButtons.OK,MessageBoxIcon.Warning);
+                return;
+            }
             Prioriteit p = new Prioriteit();
             p.Id = Convert.ToInt32(LstPrioriteiten.SelectedValue);
           
@@ -171,15 +175,23 @@ namespace MiaClient
             {
                 p.actief = false;
             }
-
+           
             if (MessageBox.Show($"Bent u zeker dat u {LstPrioriteiten.Text} wilt verwijderen?", "Prioriteit verwijderen", MessageBoxButtons.YesNo) == DialogResult.Yes)
             {
-               
-                PrioriteitManager.DeletePrioriteit(p);
-                MessageBox.Show("De Prioriteit is succesvol verwijderd", "MIA", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                PrioriteitenChanged?.Invoke(this, EventArgs.Empty);
+                if (PrioriteitManager.CheckPrioriteitInUse(p.Id))
+                {
+                    PrioriteitManager.DeactivatePrioriteit(p);
+                    PrioriteitenChanged?.Invoke(this, EventArgs.Empty);
+                }
+                else
+                {
+                    PrioriteitManager.DeletePrioriteit(p);
+                    MessageBox.Show("De Prioriteit is succesvol verwijderd", "MIA", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    PrioriteitenChanged?.Invoke(this, EventArgs.Empty);
 
+                }
             }
+                   
 
             BindLstPrioriteiten();
             ClearFields();
